LPS-139 Avoid PortalException when deleting the CommerceChannelGroup

+++ modules/apps/commerce/commerce-tax-service/src/main/java/com/liferay/commerce/tax/internal/model/listener/CommerceChannelModelListener.java
@@ -15,11 +15,13 @@
 package com.liferay.commerce.tax.internal.model.listener;
 
 import com.liferay.commerce.product.model.CommerceChannel;
+import com.liferay.commerce.product.service.CommerceChannelLocalService;
 import com.liferay.commerce.tax.service.CommerceTaxMethodLocalService;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.BaseModelListener;
+import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.ModelListener;
 
 import org.osgi.service.component.annotations.Component;
@@ -34,9 +36,16 @@ public class CommerceChannelModelListener
 
 	@Override
 	public void onBeforeRemove(CommerceChannel commerceChannel) {
+		Group group = _commerceChannelLocalService.fetchCommerceChannelGroup(
+			commerceChannel.getCommerceChannelId());
+
+		if (group == null) {
+			return;
+		}
+
 		try {
 			_commerceTaxMethodLocalService.deleteCommerceTaxMethods(
-				commerceChannel.getGroupId());
+				group.getGroupId());
 		}
 		catch (PortalException portalException) {
 			_log.error(portalException, portalException);
@@ -45,6 +54,9 @@ public class CommerceChannelModelListener
 
 	private static final Log _log = LogFactoryUtil.getLog(
 		CommerceChannelModelListener.class);
+
+	@Reference
+	private CommerceChannelLocalService _commerceChannelLocalService;
 
 	@Reference
 	private CommerceTaxMethodLocalService _commerceTaxMethodLocalService;
